# openapi/src/main/resources/hack9-reference.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Reference implementation
  license:
    name: MIT
servers:
  - url: https://hack9.levi9.com/reference
paths:
...
  /financial/invoice:
    post:
      summary: "Generate invoice"
      description: "Initiate invoice generation for all calls initiated within the given period. For each calling number and given period, one invoice needs to be generated. Each invoice has a unique ID. When invoices generation is finished, given callback URL should be called."
      operationId: createInvoice
      tags:
        - financial
      requestBody:
        description: "Invoice request"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceRequest"
        required: true
      responses:
        202:
          description: "Invoice creation is underway."
        400:
          description: "Bad parameters"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: "Error message"
...
    InvoiceRequest:
      description: "Request batch generation of invoices for all numbers and given period."
      type: object
      properties:
        start:
          type: string
          format: date-time
          description: "Start of the invoicing period."
        end:
          type: string
          format: date-time
          description: "End of the invoicing period"
        callback:
          description: "Callback URL to signal invoice completion. Callback expects a JSON of InvoiceDone"
          type: string
          format: url
      required:
        - start
        - end
        - callback
...